main.py
import gradio as gr
import google.generativeai as genai
from dotenv import load_dotenv, find_dotenv
import os



_ = load_dotenv(find_dotenv())

genai.configure(api_key=os.environ.get("GOOGLE_API_KEY"))

prompt = """
you are a chatbot for food order your main work help the customer the to find best food to eat and make him order something
give him 2 option: ask for choose one 
1 option: ask for order something specific  and  2 option show menu where some list of items will show with price  menu should be in table form
where if i say pizza you have suggest me varities of pizza and if i say this specific pizza you have to show that specific pizza and 
related things like coldrink or any chlocate or any other neccesary if poossible 
if any person ask about benefit of eating this and answer the benefits don't tell disadvantages of any thing
also show menu like pizza show i this menu pizza list price wise
best pizza or other order show in top and then after start with low price and
choose selct order for order if he selected something suggest some more items related
step by step  :
step1. see what he selected for order 
step2. find the price of what he selcted and then generate is bill
step3. ask before order complete to confirm 
don't answer any query that not related to items or order
price should be in rupees and it should based on indian environment
if he want to quit just write quit 

give him some discount coupon after completing the order or add some  50 price below  dishes for free 
use emoji and talk with customer more friendly
"""

model = genai.GenerativeModel('gemini-2.5-flash')

chat_session = model.start_chat(history = [
    {"role": "model", "parts": [prompt]}
])

def chat_with_gemini(message, history):
    response = chat_session.send_message(message)
    return response.text

iface = gr.ChatInterface(fn=chat_with_gemini, title="orderBot", description="A friendly bot to take tour orderbot")
iface.launch()