pyproject.toml
[project]
name = "burger-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "generativeai>=0.0.1",
    "google-generativeai>=0.8.5",
    "gradio>=5.39.0",
]
